feat: implemented workaround to validate regular expressions containing ?!
